+++ CHANGELOG.md
@@ -27,3 +27,4 @@
 - Quickfix to define a Slf4J logger on the class (besides Lombok @Slf4j) on unresolved "log" token
 - Reduced the size of the warning about @RAC
 - make sure static import doesn't collide with existing static imports.
+- Added action "Copy Git Coordinates to Clipboard" + menu entry in Git>